move helper and fix requires
