+++ services/importer/lib/importer/downloader.rb
@@ -15,11 +15,12 @@ require_relative './url_translator/kimono_labs'
 require_relative './unp'
 require_relative '../../../../lib/carto/http/client'
 require_relative '../../../../lib/carto/url_validator'
+require_relative '../helpers/quota_check_helpers.rb'
 
 module CartoDB
   module Importer2
     class Downloader
-
+      include CartoDB::Importer2::QuotaCheckHelpers
       extend Carto::UrlValidator
 
       # in seconds
@@ -205,8 +206,8 @@ module CartoDB
       end
 
       def set_downloaded_source_file(available_quota_in_bytes = nil)
-        raise_if_over_storage_quota(quota_requested: content_length_from(headers),
-                                    quota_available: available_quota_in_bytes.to_i,
+        raise_if_over_storage_quota(requested_quota: content_length_from(headers),
+                                    available_quota: available_quota_in_bytes.to_i,
                                     user_id: @options[:user_id])
 
         @etag           = etag_from(headers)
